Update labels layers
